rename socialLoginSuccess class to SocialLoginSuccessNewUser everywhere
