+++ app/components/Views/SocialLoginSuccessNewUser/index.styles.ts
