+++ app/components/Views/SocialLoginSuccessNewUser/index.test.tsx
