+++ app/components/Views/SocialLoginSuccessNewUser/index.tsx
