+++ app/constants/navigation/Routes.ts
@@ -117,7 +117,7 @@ const Routes = {
     ONBOARDING: 'Onboarding',
     LOGIN: 'Login',
     NAV: 'OnboardingNav',
-    SOCIAL_LOGIN_SUCCESS: 'SocialLoginSuccess',
+    SOCIAL_LOGIN_SUCCESS: 'SocialLoginSuccessNewUser',
     MANUAL_BACKUP: {
       STEP_1: 'ManualBackupStep1',
       STEP_2: 'ManualBackupStep2',
